# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PatientStatus = typeof PatientStatus[keyof typeof PatientStatus];


export const PatientStatus = {
  active: 'active',
  paused: 'paused',
} as const;

export interface Patient {
  id: string;
  name: string;
  email: string;
  initials: string;
  status: PatientStatus;
  /** @nullable */
  lastSessionAt: string | null;
  activePrescriptionCount: number;
}

export interface PatientInput {
  /** @minLength 1 */
  name: string;
  email: string;
}

export interface AngleRule {
  joint: string;
  target: number;
  min: number;
  max: number;
}

export type PrescriptionExerciseId = typeof PrescriptionExerciseId[keyof typeof PrescriptionExerciseId];


export const PrescriptionExerciseId = {
  squat: 'squat',
  'shoulder-flexion': 'shoulder-flexion',
  'sit-to-stand': 'sit-to-stand',
} as const;

export type PrescriptionStatus = typeof PrescriptionStatus[keyof typeof PrescriptionStatus];


export const PrescriptionStatus = {
  active: 'active',
  draft: 'draft',
  paused: 'paused',
} as const;

export interface Prescription {
  id: string;
  patientId: string;
  patientName: string;
  exerciseId: PrescriptionExerciseId;
  status: PrescriptionStatus;
  sets: number;
  repetitions: number;
  angleRules: AngleRule[];
  minRom: number;
  maxRom: number;
  instructions: string[];
  precautions: string[];
  updatedAt: string;
}

export type PrescriptionInputExerciseId = typeof PrescriptionInputExerciseId[keyof typeof PrescriptionInputExerciseId];


export const PrescriptionInputExerciseId = {
  squat: 'squat',
  'shoulder-flexion': 'shoulder-flexion',
  'sit-to-stand': 'sit-to-stand',
} as const;

export interface PrescriptionInput {
  patientId: string;
  exerciseId: PrescriptionInputExerciseId;
  /** @minimum 1 */
  sets: number;
  /** @minimum 1 */
  repetitions: number;
  angleRules: AngleRule[];
  minRom: number;
  maxRom: number;
  instructions: string[];
  precautions: string[];
}

export type PrescriptionUpdate = PrescriptionInput;

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  completed: 'completed',
  interrupted: 'interrupted',
} as const;

export interface Session {
  id: string;
  patientId: string;
  prescriptionId: string;
  patientName: string;
  exerciseId: string;
  startedAt: string;
  endedAt: string;
  reps: number;
  correctReps: number;
  romAchieved: number;
  durationSeconds: number;
  errors: string[];
  status: SessionStatus;
}

export type SessionInputStatus = typeof SessionInputStatus[keyof typeof SessionInputStatus];


export const SessionInputStatus = {
  completed: 'completed',
  interrupted: 'interrupted',
} as const;

export interface SessionInput {
  patientId: string;
  prescriptionId: string;
  exerciseId: string;
  startedAt: string;
  endedAt: string;
  /** @minimum 0 */
  reps: number;
  /** @minimum 0 */
  correctReps: number;
  romAchieved: number;
  durationSeconds: number;
  errors: string[];
  status: SessionInputStatus;
}

export interface DashboardSummary {
  patientCount: number;
  activePrescriptionCount: number;
  sessionsThisWeek: number;
  averageCompletion: number;
  recentSessions: Session[];
}

export type ListPrescriptionsParams = {
patientId?: string;
};

export type ListSessionsParams = {
patientId?: string;
};