# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Session } from './session';

export interface DashboardSummary {
  patientCount: number;
  activePrescriptionCount: number;
  sessionsThisWeek: number;
  averageCompletion: number;
  recentSessions: Session[];
}