# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from './patientStatus';

export interface Patient {
  id: string;
  name: string;
  email: string;
  initials: string;
  status: PatientStatus;
  /** @nullable */
  lastSessionAt: Date | null;
  activePrescriptionCount: number;
}