# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AngleRule } from './angleRule';
import type { PrescriptionExerciseId } from './prescriptionExerciseId';
import type { PrescriptionStatus } from './prescriptionStatus';

export interface Prescription {
  id: string;
  patientId: string;
  patientName: string;
  exerciseId: PrescriptionExerciseId;
  status: PrescriptionStatus;
  sets: number;
  repetitions: number;
  angleRules: AngleRule[];
  minRom: number;
  maxRom: number;
  instructions: string[];
  precautions: string[];
  updatedAt: Date;
}