# lib/api-zod/src/generated/types/prescriptionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AngleRule } from './angleRule';
import type { PrescriptionInputExerciseId } from './prescriptionInputExerciseId';

export interface PrescriptionInput {
  patientId: string;
  exerciseId: PrescriptionInputExerciseId;
  /** @minimum 1 */
  sets: number;
  /** @minimum 1 */
  repetitions: number;
  angleRules: AngleRule[];
  minRom: number;
  maxRom: number;
  instructions: string[];
  precautions: string[];
}