# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionStatus } from './sessionStatus';

export interface Session {
  id: string;
  patientId: string;
  prescriptionId: string;
  patientName: string;
  exerciseId: string;
  startedAt: Date;
  endedAt: Date;
  reps: number;
  correctReps: number;
  romAchieved: number;
  durationSeconds: number;
  errors: string[];
  status: SessionStatus;
}