# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionInputStatus } from './sessionInputStatus';

export interface SessionInput {
  patientId: string;
  prescriptionId: string;
  exerciseId: string;
  startedAt: Date;
  endedAt: Date;
  /** @minimum 0 */
  reps: number;
  /** @minimum 0 */
  correctReps: number;
  romAchieved: number;
  durationSeconds: number;
  errors: string[];
  status: SessionInputStatus;
}